+++ src/docker_healer.rs
@@ -10,12 +10,13 @@ use hyper::{Method, Response, StatusCode};
 use tokio::time::{sleep, timeout};
 use tracing::{Level, event};
 
-use crate::app_config::HealerConfig;
-use crate::container::Container;
-use crate::docker_connection::{DockerClient, DockerEndpoint};
-use crate::encoding::url_encode;
-use crate::http_client::{build_request, execute_request};
 use crate::webhook::WebHookNotifier;
+use crate::{app_config::HealerConfig, docker::client::DockerClient};
+use crate::{
+    docker::client::DockerEndpoint,
+    http_client::{build_request, execute_request},
+};
+use crate::{docker::container::Container, encoding::url_encode};
 
 pub struct DockerHealer {
     client: DockerClient,
